device-network.c: Fix crash in `pappl_dnssd_list()` with Avahi
If PAPPL testsuite runs in isolated environment, e.g. in mock,
we cannot generate Avahi client, because Avahi is not running.
This causes `_papplDNSSDInit()` to return NULL, which is later
sent into `avahi_service_browser_new()`, which uses `assert()`
for this argument and make the binary crash if the input is NULL.

The crash makes the testsuite fail during the build, which is useful sanity check before the package is built.