printer-driver.c: Fix crash when adding generic PS printer
If legacy-printer-app is run for the first time and generic PS driver
was chosen when installing the printer, the application crashes during
adding the printer.

It is because the printer structure does not have device_id set, and the
library tries to get data from already freed `printer->driver_attrs`.
Additionally, if the device id exists in the printer, it rewrites
incorrect structure (we should set device id attr in `attrs`, which we
return).

The fix adds possibility to extract device id from `printer->attrs` and if
device id is missing in `printer->attrs` and `document-format-supported`
as well, try `document-format-default` and if this attribute is missing,
return NULL.