Fix netcat hang when it send raw data to raw socket
netcat and similar tools hang when they send data to PAPPL's raw socket.
It is because these tools wait for input from the destination, and we
never end because update `activity` every time `poll()` returns
non-negative value, and the tools never send POLLHUP.

The proposed fix is to update `activity` only when we read from the
socket, and break out of the loop after 10s of inactivity.

Fixes #331