stop png read from looping endlessly
It is not safe to call png_read_end after an error.  If that function
encounters an error it will jump back to the setjump again causing
and endless loop.  Move the call to png_read_end so that it is only
performed if the png file was read successfully.